New file monitoring system
The file monitoring system now checks compares timestamps instead of
file contents.
This is way more efficient than comparing the entire file to see if it
has changed.
Since we are passing both files to the code review pipeline we actually
only need to look if the timestamp is newer than the current time minus
the scanning interval.